Update api files, design Dashboard page and handle logics

+++ src/api/category/branch/DeleteBranch.ts
@@ -1,17 +1,25 @@
 import axios from "axios";
 import envConfig from "@/src/config";
+import { ParseJSON } from "../../auth/ParseJSON";
 
 const access_token = localStorage.getItem("access_token");
 
 export const DeleteBranch = async (id: string) => {
 
     const DeleteURL = envConfig.NEXT_PUBLIC_API_ENDPOINT + `/branch/${id}`;
+
+    if (!access_token) {
+        throw new Error("No access token found");
+    }
+
+    const parseToken = ParseJSON(access_token);
+
     const config = {
         method: "delete",
         maxBodyLength: Infinity,
         url: DeleteURL,
         headers: {
-            "Authorization": `Bearer ${access_token}`,
+            "Authorization": `Bearer ${parseToken}`,
         },
     };
     try {
