+++ src/api/users/DeleteUser.ts
@@ -1,16 +1,24 @@
 import axios from 'axios';
 import envConfig from '@/src/config';
+import { ParseJSON } from '../auth/ParseJSON';
 
 const access_token = localStorage.getItem('access_token');
 
 export const DeleteUser = async (userId: string) => {
     const DeleteUserUrl = envConfig.NEXT_PUBLIC_API_ENDPOINT + `/users/${userId}`;
+
+    if (!access_token) {
+        throw new Error('No access token found');
+    }
+
+    const parseToken = ParseJSON(access_token);
+
     const config = {
         method: 'delete',
         maxBodyLength: Infinity,
         url: DeleteUserUrl,
         headers: {
-            'Authorization': `Bearer ${access_token}`,
+            'Authorization': `Bearer ${parseToken}`,
             
         }
     };
